+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-7142] Allow to sort by `non_published_note` in data structure document
+- [TD-6677] Download and upload of pending notes in Excel
 
 ## [7.2.3] 2025-03-11